requirements.txt
# Django & DRF
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
drf-spectacular>=0.27,<1.0
django-filter>=24.0,<25.0

# Database
mysqlclient>=2.2,<3.0

# Redis
django-redis>=5.4,<6.0
redis>=5.0,<6.0

# Quality
black>=24.0
flake8>=7.0
isort>=5.13
pre-commit>=3.6

# Testing
pytest>=8.0
pytest-django>=4.8
pytest-cov>=5.0
factory-boy>=3.3

# Utils
python-decouple>=3.8
structlog>=24.0
gunicorn>=22.0
